Clear caches optionally

+++ vars/render_studio.groovy
@@ -133,7 +133,7 @@ def removeClosedPRs(Map options) {
  }
 
 
-def uninstallAMDRenderStudio(String osName, Map options) {
+def uninstallAMDRenderStudio(String osName, Map options, Boolean clearCaches = true) {
     def installedProductCode = powershell(script: """(Get-WmiObject -Class Win32_Product -Filter \"Name LIKE 'AMD RenderStudio'\").IdentifyingNumber""", returnStdout: true)
 
     // TODO: compare product code of built application and installed application
@@ -144,19 +144,21 @@ def uninstallAMDRenderStudio(String osName, Map options) {
         utils.removeDir(this, osName, "C:\\Users\\%USERNAME%\\AppData\\Roaming\\AMDRenderStudio\\Storage")
         utils.removeDir(this, osName, "C:\\Users\\%USERNAME%\\Documents\\AMD RenderStudio Home")
 
-        println("[INFO] Clearing caches...")
-        for (cache in caches) {
-            if (fileExists(cache)) {
-                try {
-                    bat """
-                        del /q \"${cache}\"
-                        for /d %i in (\"${cache}\") do @rmdir /s /q "%i"
-                    """
-                    println("[INFO] Path \"${cache}\" is cleared.")
-                } catch(Exception e) {
-                    println("[ERROR] Can't clear directory")
-                    println(e.toString())
-                    println(e.getMessage())
+        if (clearCaches) {
+            println("[INFO] Clearing caches...")
+            for (cache in caches) {
+                if (fileExists(cache)) {
+                    try {
+                        bat """
+                            del /q \"${cache}\"
+                            for /d %i in (\"${cache}\") do @rmdir /s /q "%i"
+                        """
+                        println("[INFO] Path \"${cache}\" is cleared.")
+                    } catch(Exception e) {
+                        println("[ERROR] Can't clear directory")
+                        println(e.toString())
+                        println(e.getMessage())
+                    }
                 }
             }
         }
@@ -219,7 +221,7 @@ def runApplication(String appLink, String osName, Map options) {
 
 def runApplicationTests(String osName, Map options) {
     // Step 0: remove existing application
-    uninstallAMDRenderStudio(osName, options)
+    uninstallAMDRenderStudio(osName, options, false)
 
     // Step 1: install app
     installAMDRenderStudio(osName, options)
@@ -242,7 +244,7 @@ def runApplicationTests(String osName, Map options) {
     runApplication(appLink, osName, options)
 
     // Step 5: try to uninstall app (should be unsuccessfull, issue on Render Studio side)
-    uninstallAMDRenderStudio(osName, options)
+    uninstallAMDRenderStudio(osName, options, false)
 
     // Step 6: close app window
     try {
@@ -258,7 +260,7 @@ def runApplicationTests(String osName, Map options) {
     // }
 
     // Step 8: uninstall app
-    uninstallAMDRenderStudio(osName, options)
+    uninstallAMDRenderStudio(osName, options, false)
 
     // Step 9: check processes after uninstalling
     Boolean processFound = utils.isProcessExists(this, "AMD RenderStudio", osName, options)
@@ -375,7 +377,7 @@ def prepareAMDRenderStudio(String osName, Map options, String clientType, int cl
 
         withNotifications(title: options["stageName"], options: options, configuration: NotificationConfiguration.INSTALL_PLUGIN) {
             timeout(time: "10", unit: "MINUTES") {
-                uninstallAMDRenderStudio(osName, options)
+                uninstallAMDRenderStudio(osName, options, true)
                 installAMDRenderStudio(osName, options)
 
                 // open application once to generate Storage files
